feat: Extracted 4 prompt-lifecycle handlers (prompt, steer, follow_up,…
- src/extension/handlers/prompt-handlers.ts
- src/extension/handlers/__tests__/prompt-handlers.test.ts
- src/extension/message-dispatch.ts
- src/extension/message-dispatch.test.ts

GSD-Task: S01/T04
